+++ releasenotes.md
@@ -1,4 +1,3 @@
 Whats New:
 ----------------------
-Multiple Documents per ingredients added. You can store MSDS, COFA or any other files related to ingredient
-Various bug fixes and improvements
+Performance improvements and bug fixes